Struct rusoto_lex_runtime::GenericAttachment

pub struct GenericAttachment {
    pub attachment_link_url: Option<String>,
    pub buttons: Option<Vec<Button>>,
    pub image_url: Option<String>,
    pub sub_title: Option<String>,
    pub title: Option<String>,
}

Represents an option rendered to the user when a prompt is shown. It could be an image, a button, a link, or text.

Fields
attachment_link_url: Option<String>

The URL of an attachment to the response card.

buttons: Option<Vec<Button>>

The list of options to show to the user.

image_url: Option<String>

The URL of an image that is displayed to the user.

sub_title: Option<String>

The subtitle shown below the title.

title: Option<String>

The title of the option.