Crate rusoto_lex_runtime

Amazon Lex provides both build and runtime endpoints. Each endpoint provides a set of operations (API). Your conversational bot uses the runtime API to understand user utterances (user input text or voice). For example, suppose a user says "I want pizza", your bot sends this input to Amazon Lex using the runtime API. Amazon Lex recognizes that the user request is for the OrderPizza intent (one of the intents defined in the bot). Then Amazon Lex engages in user conversation on behalf of the bot to elicit required information (slot values, such as pizza size and crust type), and then performs fulfillment activity (that you configured when you created the bot). You use the build-time API to create and manage your Amazon Lex bot. For a list of build-time operations, see the build-time API, .

If you're using the service, you're probably looking for LexRuntimeClient and LexRuntime.

Structs

Button
Represents an option to be shown on the client platform (Facebook, Slack, etc.)

GenericAttachment
Represents an option rendered to the user when a prompt is shown. It could be an image, a button, a link, or text.

LexRuntimeClient
A client for the Amazon Lex Runtime Service API.

PostContentRequest

PostContentResponse

PostTextRequest

PostTextResponse

ResponseCard
If you configure a response card when creating your bots, Amazon Lex substitutes the session attributes and slot values that are available, and then returns it. The response card can also come from a Lambda function ( dialogCodeHook and fulfillmentActivity on an intent).

Enums

PostContentError
Errors returned by PostContent

PostTextError
Errors returned by PostText

Traits

LexRuntime
Trait representing the capabilities of the Amazon Lex Runtime Service API. Amazon Lex Runtime Service clients implement this trait.