Struct rusoto_lex_runtime::PostContentResponse

pub struct PostContentResponse {
    pub audio_stream: Option<Vec<u8>>,
    pub content_type: Option<String>,
    pub dialog_state: Option<String>,
    pub input_transcript: Option<String>,
    pub intent_name: Option<String>,
    pub message: Option<String>,
    pub session_attributes: Option<String>,
    pub slot_to_elicit: Option<String>,
    pub slots: Option<String>,
}

Fields
audio_stream: Option<Vec<u8>>

The prompt (or statement) to convey to the user. This is based on the bot configuration and context. For example, if Amazon Lex did not understand the user intent, it sends the clarificationPrompt configured for the bot. If the intent requires confirmation before taking the fulfillment action, it sends the confirmationPrompt. Another example: Suppose that the Lambda function successfully fulfilled the intent, and sent a message to convey to the user. Then Amazon Lex sends that message in the response.

content_type: Option<String>

Content type as specified in the Accept HTTP header in the request.

dialog_state: Option<String>

Identifies the current state of the user interaction. Amazon Lex returns one of the following values as dialogState. The client can optionally use this information to customize the user interface.

ElicitIntent – Amazon Lex wants to elicit the user's intent. Consider the following examples:

For example, a user might utter an intent ("I want to order a pizza"). If Amazon Lex cannot infer the user intent from this utterance, it will return this dialog state.

ConfirmIntent – Amazon Lex is expecting a "yes" or "no" response.

For example, Amazon Lex wants user confirmation before fulfilling an intent. Instead of a simple "yes" or "no" response, a user might respond with additional information. For example, "yes, but make it a thick crust pizza" or "no, I want to order a drink." Amazon Lex can process such additional information (in these examples, update the crust type slot or change the intent from OrderPizza to OrderDrink).

ElicitSlot – Amazon Lex is expecting the value of a slot for the current intent.

For example, suppose that in the response Amazon Lex sends this message: "What size pizza would you like?". A user might reply with the slot value (e.g., "medium"). The user might also provide additional information in the response (e.g., "medium thick crust pizza"). Amazon Lex can process such additional information appropriately.

Fulfilled – Conveys that the Lambda function has successfully fulfilled the intent.

ReadyForFulfillment – Conveys that the client has to fullfill the request.

Failed – Conveys that the conversation with the user failed.

This can happen for various reasons, including that the user does not provide an appropriate response to prompts from the service (you can configure how many times Amazon Lex can prompt a user for specific information), or if the Lambda function fails to fulfill the intent.

input_transcript: Option<String>

Transcript of the voice input to the operation.

intent_name: Option<String>

Current user intent that Amazon Lex is aware of.

message: Option<String>

Message to convey to the user. It can come from the bot's configuration or a code hook (Lambda function). If the current intent is not configured with a code hook or if the code hook returned Delegate as the dialogAction.type in its response, then Amazon Lex decides the next course of action and selects an appropriate message from the bot configuration based on the current user interaction context. For example, if Amazon Lex is not able to understand the user input, it uses a clarification prompt message (For more information, see the Error Handling section in the Amazon Lex console). Another example: if the intent requires confirmation before fulfillment, then Amazon Lex uses the confirmation prompt message in the intent configuration. If the code hook returns a message, Amazon Lex passes it as-is in its response to the client.

session_attributes: Option<String>

Map of key/value pairs representing the session-specific context information.

slot_to_elicit: Option<String>

If the dialogState value is ElicitSlot, returns the name of the slot for which Amazon Lex is eliciting a value.

slots: Option<String>

Map of zero or more intent slots (name/value pairs) Amazon Lex detected from the user input during the conversation.