Struct rusoto_lex_runtime::PostTextResponse

pub struct PostTextResponse {
    pub dialog_state: Option<String>,
    pub intent_name: Option<String>,
    pub message: Option<String>,
    pub response_card: Option<ResponseCard>,
    pub session_attributes: Option<HashMap<String, String>>,
    pub slot_to_elicit: Option<String>,
    pub slots: Option<HashMap<String, String>>,
}

Fields
dialog_state: Option<String>

Identifies the current state of the user interaction. Amazon Lex returns one of the following values as dialogState. The client can optionally use this information to customize the user interface.

ElicitIntent – Amazon Lex wants to elicit user intent.

For example, a user might utter an intent ("I want to order a pizza"). If Amazon Lex cannot infer the user intent from this utterance, it will return this dialogState.

ConfirmIntent – Amazon Lex is expecting a "yes" or "no" response.

For example, Amazon Lex wants user confirmation before fulfilling an intent.

Instead of a simple "yes" or "no," a user might respond with additional information. For example, "yes, but make it thick crust pizza" or "no, I want to order a drink". Amazon Lex can process such additional information (in these examples, update the crust type slot value, or change intent from OrderPizza to OrderDrink).

ElicitSlot – Amazon Lex is expecting a slot value for the current intent.

For example, suppose that in the response Amazon Lex sends this message: "What size pizza would you like?". A user might reply with the slot value (e.g., "medium"). The user might also provide additional information in the response (e.g., "medium thick crust pizza"). Amazon Lex can process such additional information appropriately.

Fulfilled – Conveys that the Lambda function configured for the intent has successfully fulfilled the intent.

ReadyForFulfillment – Conveys that the client has to fulfill the intent.

Failed – Conveys that the conversation with the user failed.

This can happen for various reasons including that the user did not provide an appropriate response to prompts from the service (you can configure how many times Amazon Lex can prompt a user for specific information), or the Lambda function failed to fulfill the intent.

intent_name: Option<String>

The current user intent that Amazon Lex is aware of.

message: Option<String>

A message to convey to the user. It can come from the bot's configuration or a code hook (Lambda function). If the current intent is not configured with a code hook or the code hook returned Delegate as the dialogAction.type in its response, then Amazon Lex decides the next course of action and selects an appropriate message from the bot configuration based on the current user interaction context. For example, if Amazon Lex is not able to understand the user input, it uses a clarification prompt message (for more information, see the Error Handling section in the Amazon Lex console). Another example: if the intent requires confirmation before fulfillment, then Amazon Lex uses the confirmation prompt message in the intent configuration. If the code hook returns a message, Amazon Lex passes it as-is in its response to the client.

response_card: Option<ResponseCard>

Represents the options that the user has to respond to the current prompt. Response Card can come from the bot configuration (in the Amazon Lex console, choose the settings button next to a slot) or from a code hook (Lambda function).

session_attributes: Option<HashMap<String, String>>

A map of key-value pairs representing the session-specific context information.

slot_to_elicit: Option<String>

If the dialogState value is ElicitSlot, returns the name of the slot for which Amazon Lex is eliciting a value.

slots: Option<HashMap<String, String>>

The intent slots (name/value pairs) that Amazon Lex detected so far from the user input in the conversation.